Integrate log equilibrium into LTE composition
